shifted code from MainActivity to EventsFragment
