+++ app/build.gradle
@@ -45,6 +45,7 @@ dependencies {
 
     //added deps for hilt.
     implementation "com.google.dagger:hilt-android:2.38.1"
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     kapt "com.google.dagger:hilt-compiler:2.38.1"
 
     testImplementation 'junit:junit:4.+'
