+++ app/src/main/java/com/iiitlucknow/festify/MainActivity.kt
@@ -13,35 +13,11 @@ import dagger.hilt.android.AndroidEntryPoint
 @AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
-    private lateinit var sliderAdapter: sliderAdapter
-    private var images: MutableList<Int> = mutableListOf(
-        R.drawable.android_img,
-        R.drawable.web_development_img,
-        R.drawable.cybersecurity_img,
-        R.drawable.ai_img,
-        R.drawable.designing_img
-    )
-    private var fields: MutableList<recyclerItem> = mutableListOf(
-        recyclerItem(R.drawable.android, R.string.android),
-        recyclerItem(R.drawable.web_development, R.string.web),
-        recyclerItem(R.drawable.cyber_security, R.string.cyber_sec),
-        recyclerItem(R.drawable.ai, R.string.ai),
-        recyclerItem(R.drawable.design, R.string.design)
-    )
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
         supportActionBar?.hide()
-        sliderAdapter = sliderAdapter(images)
-        binding.apply {
-            slide.setSliderAdapter(sliderAdapter)
-            slide.setIndicatorAnimation(IndicatorAnimationType.WORM)
-            slide.setSliderTransformAnimation(SliderAnimations.DEPTHTRANSFORMATION)
-            slide.startAutoCycle()
-            recycler.adapter = recyclerAdapter(fields)
-            recycler.setHasFixedSize(true)
-        }
     }
 }
