+++ app/src/main/res/layout/activity_main.xml
@@ -7,39 +7,10 @@
     tools:context=".MainActivity">
 
 
-    <com.smarteist.autoimageslider.SliderView
-        android:id="@+id/slide"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:sliderAnimationDuration="600"
-        app:sliderAutoCycleDirection="back_and_forth"
-        app:sliderAutoCycleEnabled="true"
-        app:sliderIndicatorAnimationDuration="600"
-        app:sliderIndicatorGravity="center_horizontal|bottom"
-        app:sliderIndicatorMargin="15dp"
-        app:sliderIndicatorOrientation="horizontal"
-        app:sliderIndicatorPadding="3dp"
-        app:sliderIndicatorRadius="2dp"
-        app:sliderIndicatorSelectedColor="#5A5A5A"
-        app:sliderIndicatorUnselectedColor="#FFF"
-        app:sliderScrollTimeInSec="2"
-        app:sliderStartAutoCycle="true" />
-
-    <androidx.recyclerview.widget.RecyclerView
-
-        android:id="@+id/recycler"
-        android:layout_width="match_parent"
-        android:layout_height="530dp"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/slide"
-        tools:listitem="@layout/recycler_item_layout" />
+   <androidx.fragment.app.FragmentContainerView
+       android:name="com.iiitlucknow.festify.EventsFragment"
+       android:layout_width="match_parent"
+       android:layout_height="match_parent"
+       android:id="@+id/new_fragment"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
